# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIKARIR API",
        "description": "Dokumentasi API untuk aplikasi Sikarir — Sistem Informasi Karir & Pelatihan BLK",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/kelas-pelatihan": {
            "get": {
                "tags": [
                    "Kelas Pelatihan"
                ],
                "summary": "Ambil semua data kelas pelatihan (paginate 15)",
                "operationId": "9a0966e7b40dfc1117ad16ebef88bddc",
                "responses": {
                    "200": {
                        "description": "Daftar kelas pelatihan berhasil diambil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Kelas Pelatihan"
                ],
                "summary": "Daftarkan peserta ke kelas pelatihan",
                "operationId": "a510dd498e0e89fa5d4b83deb526628b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_peserta",
                                    "id_jadwal"
                                ],
                                "properties": {
                                    "id_peserta": {
                                        "description": "ID peserta yang akan didaftarkan",
                                        "type": "integer"
                                    },
                                    "id_jadwal": {
                                        "description": "ID jadwal pelatihan yang dituju",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Peserta berhasil didaftarkan ke kelas"
                    },
                    "422": {
                        "description": "Jadwal tidak tersedia / kuota penuh / peserta sudah terdaftar"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/kelas-pelatihan/{id}": {
            "get": {
                "tags": [
                    "Kelas Pelatihan"
                ],
                "summary": "Ambil detail kelas pelatihan berdasarkan ID",
                "operationId": "858da82f0cce434b673b3003b485e1db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail kelas pelatihan"
                    },
                    "404": {
                        "description": "Kelas tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Kelas Pelatihan"
                ],
                "summary": "Update data kelas pelatihan",
                "operationId": "ca85a06e32362c451a7ed41049144eef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id_peserta": {
                                        "type": "integer"
                                    },
                                    "id_jadwal": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kelas berhasil diupdate"
                    },
                    "404": {
                        "description": "Kelas tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Kelas Pelatihan"
                ],
                "summary": "Hapus data kelas pelatihan",
                "operationId": "5ded9978ab21e94d161821ff3bef435a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kelas berhasil dihapus"
                    },
                    "404": {
                        "description": "Kelas tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Masukkan token yang diperoleh dari endpoint /api/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}